requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
discord.py==1.3.4
idna==2.10
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
mmh3==2.5.1
more-itertools==8.4.0
multidict==4.7.6
portend==2.6
pottery==0.71
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
six==1.15.0
tempora==3.0.0
websockets==8.1
yarl==1.4.2
zc.lockfile==2.0
